+++ spec/spec_helper.rb
@@ -64,9 +64,7 @@ end
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }
 
 def inspect_source(cop, source)
-  ast, comments, tokens, _ = Rubocop::CLI.parse('(string)') do |source_buffer|
-    source_buffer.source = source.join($RS)
-  end
+  ast, comments, tokens, _ = Rubocop::CLI.parse('(string)', source.join($RS))
   cop.inspect(source, tokens, ast, comments)
 end