+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * New cop `BlockNesting` checks for excessive block nesting
+* New cop `CaseEquality` checks for explicit use of `===`
 * Introduced formatter feature, enables custom formatted output and multiple outputs.
 * Added `-f`/`--format` option to specify formatter.
 * Added `-o`/`--out` option to specify output file for each formatter.