+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Bug fixes
 
 * Correct calculation of whether a modifier version of a conditional statement will fit.
+* Fix an error in `MultilineIfThen` cop that occurred in some special cases.
 
 ## 0.8.1 (05/30/2013)